requirements.txt
streamlit
numpy
inference-sdk
Pillow